Compute volume under 2D gaussian

Compute volume under 2D gaussian

get_volume_gaussian_2D(X_sig, Y_sig)

Arguments

X_sig
numeric value(s) of the x-axis spread (sigma)

Y_sig
numeric value(s) of the y-axis spread (sigma)

Value

Numeric value(s) indicating the computed volume(s)

Details

Volume under the 2D gaussian is computed as: 2 * pi * sqrt(abs(X_sig)) * sqrt(abs(Y_sig))

Numeric vectors can be supplied to X_sig and Y_sig. If vectors of length greater than 1 are given, the function computes volume for each sequential pair of X_sig, Y_sig values. The lengths of these supplied vectors must be identical.

Author

Vikram B. Baliga

Examples

library(gaussplotR)

get_volume_gaussian_2D(5, 3) #24.33467

#> [1] 24.33467